A couple changes to make it compile with the latest (buggy) Windows SDK.  It should now actually compile with the latest versions, without any modification.
Removed QEdit.h dependency, as it's mysteriously missing in the latest Windows SDKs, and even if you have it, it depends on dxtrans.h, which is mysteriously missing from the latest DirectX SDK.  To do this, I removed the inactive ISampleGrabber code for the internal audio player, which no script uses anyways.

Also updated winver to 0x600.  This unfortunately makes it possible to accidentally make a build that won't run under pre-Vista windows versions, but endpointvolume won't compile without it, unless I maintain a list of definitions to stop several blocks from being compiled.